requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
groq==0.30.0
newspaper3k==0.2.8
python-decouple==3.8
duckduckgo-search==8.1.1
lxml[html_clean]==5.3.0
requests==2.32.3
beautifulsoup4==4.12.3

